Add data checking and error messages to uploader

+++ backend/src/model/state.rs
@@ -3,6 +3,6 @@ use sqlx::FromRow;
 
 #[derive(FromRow, Deserialize, Serialize)]
 pub struct State {
-    pub id: i16,
+    pub id: i32,
     pub name: String,
 }
